+++ backpacker/php-files/db_config.php
@@ -1,0 +1,7 @@
+<?php
+
+define('DB_USER', "root");        // db user
+define('DB_PASSWORD', "111111"); // db password (mention your db password here)
+define('DB_DATABASE', "test"); // database name
+define('DB_SERVER', "localhost"); // db server
+?>
