+++ backpacker/php-files/input_get_img.php
@@ -1,0 +1,4 @@
+<form action="get_img.php" method="get">
+  name : <input type="text" name="name" /></br>
+  <input type="submit" value="GET" />
+</form>
