+++ backpacker/php-files/input_login.php
@@ -1,0 +1,5 @@
+<form action="login.php" method="post">
+  email : <input type="text" name="email" /></br>
+  pw : <input type="text" name="pw" /></br>
+  <input type="submit" value="POST" />
+</form>
