+++ backpacker/php-files/input_save_img.php
@@ -1,0 +1,5 @@
+<form action="save_img.php" method="post">
+  image name : <input type="text" name="image_name" /></br>
+  image string : <input type="text" name="encoded_string" /></br>
+  <input type="submit" value="POST" />
+</form>
