--- input_join.php
@@ -1,6 +1,0 @@
-<form action="join.php" method="post">
-  email : <input type="text" name="email" /></br>
-  pw : <input type="text" name="pw" /></br>
-  pw_: <input type="text" name="pw_" /></br>
-  <input type="submit" value="POST" />
-</form>
